fix(callblasting): seconds sin redondeo (float) en custom y summary #8
- CalculateDurationCustom: reemplaza ceil().cast(Int32) por round(3) — valor exacto
- CBSummaryGroup.seconds y CBSummaryGeneral.total_seconds: int → float
- Tests actualizados con pytest.approx para valores float

+++ src/modules/process/domain/models/summary.py
@@ -38,14 +38,14 @@ class CampaignSummary(BaseModel):
 class CBSummaryGroup(BaseModel):
     operator: str
     total: int
-    seconds: int
+    seconds: float
     credits: float
     unit_value: float
 
 
 class CBSummaryGeneral(BaseModel):
     total_records: int
-    total_seconds: int
+    total_seconds: float
     total_credits: float
     total_excluded: int
